添加icache

+++ npc/include/memory/host.h
@@ -14,9 +14,9 @@ static inline word_t host_read(void *addr, int len) {
 
 static inline void host_write(void *addr, int len, word_t data) {
   switch (len) {
-    case 1: *(uint8_t  *)addr = data; return;
-    case 2: *(uint16_t *)addr = data; return;
-    case 4: *(uint32_t *)addr = data; return;
+    case 1: *(uint8_t  *)addr = (uint8_t)data; return;
+    case 2: *(uint16_t *)addr = (uint16_t)data; return;
+    case 4: *(uint32_t *)addr = (uint32_t)data; return;
     default: assert(0);
   }
 }
